Cleanup of code
Added JavaDoc to Lists.java
Cleaned up code formatting, removed used of the i variable for the for
loop (replaced with counter).
Cleaned up console output, easier to tell which values are before and
after the use of ListIterator now.
Commented out package because of the lack of package-info and the file
is not within the package folder.